requirements.txt
python-telegram-bot[job-queue]>=21.0
matplotlib>=3.8
requests>=2.31
